clear init and introduce modules, so it will be easier to recieve more builtin functions
